Add clippy

+++ tasks/abc138c/src/main.rs
@@ -8,7 +8,7 @@ fn main() {
     v.sort();
     let answer: f64 = v
         .iter()
-        .fold(v[0] as f64, |acc, &x| (acc + x as f64) / 2 as f64);
+        .fold(v[0] as f64, |acc, &x| (acc + x as f64) / 2_f64);
 
     println!("{}", answer);
 }
